+++ wizard/app.py
@@ -7,6 +7,23 @@ from views.side_buttons import render_side_buttons
 from views.splash import render_splash
 from views.wizard import render_wizard
 
+
+def init_state():
+
+    if Metadata not in st.session_state:
+        st.session_state[Metadata] = Metadata()
+
+    if mlc.Dataset not in st.session_state:
+        st.session_state[mlc.Dataset] = None
+
+    if CurrentStep not in st.session_state:
+        st.session_state[CurrentStep] = "start"
+
+def set_form_step(action, step=None):
+    """Maintains the user's location within the wizard."""
+    if action == "Jump" and step is not None:
+        st.session_state[CurrentStep] = step
+
 init_state()
 
 st.set_page_config(page_title="Croissant Wizard", layout="wide")
@@ -24,3 +41,4 @@ elif st.session_state[CurrentStep] == CurrentStep.editor:
         render_wizard()
 else:
     st.warning("invalid unhandled app state")
+
