+++ wizard/core/state.py
@@ -6,22 +6,11 @@ In the future, this could be the serialization format between front and back.
 import dataclasses
 from typing import Type, TypeVar
 
-import streamlit as st
-
 import mlcroissant as mlc
 
 T = TypeVar('T', bound='Parent')
 
-def init_state():
 
-    if Metadata not in st.session_state:
-        st.session_state[Metadata] = Metadata()
-
-    if mlc.Dataset not in st.session_state:
-        st.session_state[mlc.Dataset] = None
-
-    if CurrentStep not in st.session_state:
-        st.session_state[CurrentStep] = "start"
 
 
 class CurrentStep:
