--- wizard/feat
@@ -1,1 +1,0 @@
-put local files in .cache/croissant
