+++ wizard/views/side_buttons.py
@@ -1,5 +1,5 @@
+from app import set_form_step
 from core.state import CurrentStep
-from core.state import set_form_step
 import streamlit as st
